% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks-functions.R
\name{entrezid2url}
\alias{entrezid2url}
\title{URL's from entrez identifiers}
\usage{
entrezid2url(id)
}
\arguments{
\item{id}{ENTREZ identifiers}
}
\value{
The corresponding URL's in the database
}
\description{
It makes the url's from entrez identifiers
}
\concept{URL generation}
