% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks-data.R
\docType{data}
\name{geod19804}
\alias{geod19804}
\title{geod19804}
\format{
ExpressionSet
}
\description{
ExpressionSet \code{geod19804}
}
\examples{
data(geod19804,package="carlosmarrero")
}
\keyword{datasets}
