% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks-data.R
\docType{data}
\name{PRJNA693328}
\alias{PRJNA693328}
\title{PRJNA693328}
\format{
SummarizedExperiment
}
\description{
SummarizedExperiment \code{PRJNA693328}
}
\examples{
data(PRJNA693328,package="carlosmarrero")
}
\keyword{datasets}
