% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks-functions.R
\name{TrapezoidRule}
\alias{TrapezoidRule}
\title{Trapezoid rule for numerical integration}
\usage{
TrapezoidRule(x, y)
}
\arguments{
\item{x}{Abscisas}

\item{y}{Function values at \code{x}}
}
\description{
An implementation of the trapezoid rule for numerical integration
}
